hub_ai/pyproject.toml
[tool.poetry]
name = "price-benchmarking"  # No spaces allowed
version = "0.1.0"
authors = ["Your Name <your.email@example.com>"] 
description = "Web scraping and data processing for price benchmarking"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crawl4ai = "==0.6.3"
pandas = "==2.2.2"
loguru = "==0.7.3"
pydantic = "==2.11.5"
boto3 = "==1.38.30"
botocore = "==1.38.30"
snowflake-connector-python = { version = "==3.15.0", extras = ["pandas"] }
aiohttp = ">=3.11.11"
requests = ">=2.31.0"
openai = ">=1.30.1,<2.0.0"
scikit-learn = ">=1.3.2,<2.0.0"
psycopg2-binary = ">=2.9.9,<3.0.0"
fastapi = "^0.111.0"
uvicorn = "^0.34.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.4"
tqdm = "^4.66.4"
tenacity = "^9.1.2"
snowflake-snowpark-python = "^1.34.0"

[tool.poetry.group.dev.dependencies]
async-timeout = "4.0.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"